Add tox.
I'm still a little suspicious of venvs-within-venvs and a billion copies of pandas but
isolation *is* good and I guess this is still the standard for python things and there
are a lot of plugins and things. And I still need a Make-ish framework which is how I got
here in the first place so let's at least try it out.